# TST: Speed up travis a bit