--- a/.travis.yml
+++ b/.travis.yml
@@ -46,7 +46,7 @@ install:
   # Customise the testing environment
   # ---------------------------------
   - conda config --add channels conda-forge
-  - PACKAGES="$PACKAGES cython pillow pytest pytest-xdist pep8 proj4=4.* pyshp shapely six requests pyepsg owslib"
+  - PACKAGES="$PACKAGES cython pillow pytest pytest-xdist filelock pep8 proj4=4.* pyshp shapely six requests pyepsg owslib"
   - conda install --quiet $PACKAGES
 
   # Conda debug