--- a/INSTALL
+++ b/INSTALL
@@ -114,6 +114,9 @@ Testing Dependencies
 ~~~~~~~~~~~~~~~~~~~~
 These packages are required for the full Cartopy test suite to run.
 
+**filelock** (https://filelock.readthedocs.io/)
+    A platform independent file lock for Python.
+
 **mock** 1.0.1 (https://pypi.python.org/pypi/mock/)
     Python mocking and patching package for testing. Note that this package
     is only required to support the Cartopy unit tests.