--- a/requirements/tests.txt
+++ b/requirements/tests.txt
@@ -1,2 +1,3 @@
+filelock
 mock>=1.0.1
 pytest>=3.0.0