@QuLogic 2018-09-26
Because we re-use filenames for tests, I added a lock using `filelock` (since it's used by Iris already) when doing I/O for the image tests. This prevents one process from overwriting the file a different process is trying to read.